refactor: Enum타입 Command 패키지 변경 + 비교시 == 연산자 적용
